feat: member 패키지 수정
헥사고날 아키텍처에 맞게 adapter와 application으로 변경
