+++ members/src/main/java/com/kor/syh/members/adapter/in/web/FindMemberRequest.java
@@ -1,0 +1,19 @@
+package com.kor.syh.members.adapter.in.web;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Getter
+public class FindMemberRequest {
+
+	@NotBlank
+	private String loginId;
+
+	@NotBlank
+	private String password;
+}
