--- members/src/main/java/com/kor/syh/members/application/in/web/LoginMemberRequest.java
@@ -1,14 +1,0 @@
-package com.kor.syh.members.application.in.web;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Getter
-public class LoginMemberRequest {
-	private String loginId;
-	private String password;
-}
