+++ members/src/main/java/com/kor/syh/members/application/port/in/auth/LoginMemberUseCase.java
@@ -1,0 +1,5 @@
+package com.kor.syh.members.application.port.in.auth;
+
+public interface LoginMemberUseCase {
+	String login(String loginId,String password);
+}
