+++ members/src/main/java/com/kor/syh/members/application/port/in/member/FindMemberUseCase.java
@@ -1,0 +1,5 @@
+package com.kor.syh.members.application.port.in.member;
+
+public interface FindMemberUseCase {
+	FindMemberResponse find(String loginId, String password);
+}
