+++ members/src/main/java/com/kor/syh/members/application/port/in/member/RegisterMemberUseCase.java
@@ -1,0 +1,5 @@
+package com.kor.syh.members.application.port.in.member;
+
+public interface RegisterMemberUseCase {
+	void register(RegisterMemberCommand command);
+}
