--- members/src/main/java/com/kor/syh/members/port/in/auth/LoginMemberUseCase.java
@@ -1,5 +1,0 @@
-package com.kor.syh.members.port.in.auth;
-
-public interface LoginMemberUseCase {
-	String login(String loginId,String password);
-}
