--- members/src/main/java/com/kor/syh/members/port/in/member/FindMemberUseCase.java
@@ -1,5 +1,0 @@
-package com.kor.syh.members.port.in.member;
-
-public interface FindMemberUseCase {
-	FindMemberResponse find(String loginId, String password);
-}
