--- members/src/main/java/com/kor/syh/members/port/in/member/RegisterMemberUseCase.java
@@ -1,5 +1,0 @@
-package com.kor.syh.members.port.in.member;
-
-public interface RegisterMemberUseCase {
-	void register(RegisterMemberCommand command);
-}
